feat(macos): implement high-accuracy latency and fix streaming stability
- Replicated RTT-aware latency metrics from the browser to the macOS native client.
- Implemented /timez endpoint and ping-pong mechanism on the macOS host server.
- Migrated native client pipeline to use int64 millisecond timestamps for all frame tracking.
- Decoupled video presentation timing from absolute server timestamps in the macOS renderer
  to eliminate input lag caused by clock skew, using local monotonic sequencing instead.
- Enforced immediate display of decoded frames using kCMSampleAttachmentKey_DisplayImmediately.
- Added server and client-side guards to ignore redundant resize/config requests, preventing
  aggressive encoder/agent restart loops and connection drops.
- Verified fix across macOS native client and browser environments.

+++ client/native_types.go
@@ -60,7 +60,7 @@ type NativeWindowLifecycle struct {
 type NativeFramePresented struct {
 	Width                        int
 	Height                       int
-	PacketTimestamp              uint32
+	PacketTimestamp              int64
 	FirstPacketSequenceNumber    uint16
 	Brightness                   int
 	ProbeMarker                  int
@@ -75,7 +75,7 @@ type NativeFramePresented struct {
 }
 
 type LatencyBreakdown struct {
-	PacketTimestamp              uint32 `json:"packetTimestamp"`
+	PacketTimestamp              int64  `json:"packetTimestamp"`
 	FirstPacketSequenceNumber    uint16 `json:"firstPacketSequenceNumber,omitempty"`
 	Brightness                   int    `json:"brightness"`
 	ProbeMarker                  int    `json:"probeMarker,omitempty"`
@@ -120,7 +120,7 @@ type TimedVideoFrameHandler interface {
 	HandleVideoFrameWithTiming(
 		codec string,
 		frame []byte,
-		packetTimestamp uint32,
+		packetTimestamp int64,
 		firstPacketSequenceNumber uint16,
 		firstDecryptedPacketQueuedAt int64,
 		firstRemotePacketAt int64,
